+++ src/buildDevLoaderFile.js
@@ -8,16 +8,20 @@ const relativizeCssUrls = require('./relativizeCssUrls.js');
 // Note: module files are to be copied or symlinked separately
 
 async function buildDevLoaderFile(settings, type) {
+	const assetList = await Promise.all(Object.keys(settings.collations).map(collationName => {
+		return each(settings, collationName, settings.collations[collationName], type);
+	}));
+	return assetList.reduce((acc, curr) => acc.concat(curr), []);
+}
+
+async function each(settings, name, modules, type) {
 	let {
 		base,
-		modules,
 		buildDevDir,
-		buildCssFile,
-		buildJsFile,
 		buildAssetsDir,
 	} = settings;
 	
-	let outputFile = type === 'css' ? buildCssFile : type === 'js' ? buildJsFile : null;
+	let outputFile = `${name}.${type}`;
 
 	let outputPath = path.resolve(path.join(base, buildDevDir, outputFile));
 	let mirrorDir = path.resolve(path.join(base, buildDevDir, buildAssetsDir));